ci: do not validate single commit message (#111)
# Description

Disables the check to validate the commit message instead of the PR
title. This repository uses the whole PR as commit message, so it makes
no sense to validate the commit of the feature branch. This prevents the
action from failing if there is one commit only.

# Verification

Not needed. CI change only.

# Checklist

- [ ] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation
- [ ] I have used pre-commit hook to update the Terraform documentation