fix: set token hop_limit to 1 (#134)
# Description

This PR reduces the hop_limit of the IMDS token to 1. This is a bastion
host used for port forwarding only. There shouldn't be any jobs running.

Reason: The token should not leave the machine and we do not use Docker
here (needs a hop_limit of at least 2).

# Verification

No verification done.

# Checklist

- [x] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation