feat: add `var.ami_id` to fix the AMI used for the Bastion (#199)
# Description

The infrastructure should be immutable, i.e. applying the same
infrastructure every week shouldn't change anything. At the moment the
AMI id changes when a new AMI is published. This might introduce
problems as the image is changed. It also increases the elapsed time of
your workflows/pipelines by roughly 4 minutes.

This PR adds an optional `var.ami_id`. In case it is filled, this AMI is
used for the Bastion EC2 instance. The `var.ami_name_filter` is now
deprecated and will be removed with version 3.0.0 of the module.

I recommend using this new parameter. But make sure that you have an
update procedure in place to update the AMI id regularly.

# Verification

Apply the `simple` example --> Bastion host is setup and boots

# Checklist

- [x] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation