chore: use PAT to auto-approve Renovate PRs (#204)
Uses the new PAT `secrets.PAT_FOR_PR_AUTO_APPROVAL` to approve Renovate
PRs. Temporary created access tokens via GitHub App do not work here as
the app is not a valid approver.

To use the auto-merge feature, make sure that the owner of the PAT is in
the list of approvers.