chore!: update the default image via Renovate (#336)
# Description

So far the AMI id was calculated on every `terraform apply` and changed
frequently whenever a new AMI was published. This is against the concept
of immutable infrastructure. The AMI should not change automatically.

This PR enables Renovate to update the AMI via PR.

# Migrations required

The `var.ami_name_filter` is removed. The AMI id is set internally or
can be provided with `var.ami_id` if you want to use (and update) your
own AMI

# Checklist

- [x] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code