feat!: import the IAM connection role and security group
Roles are typically interconnected. So it is better to create the role outside the module and import its name.

Usually security groups are related to each other and it is better to create them outside a module in a central place. Therefor we no longer create the security group inside the module.