feat!: create security group outside the module (#382)
# Description

Roles are typically interconnected. So it is better to create the role
outside the module and import its name.

Usually security groups are related to each other and it is better to
create them outside a module in a central place. Therefor we no longer
create the security group inside the module.

# Migrations required

Create the security group and IAM role outside the module and pass the
values.

```hcl
resource "aws_security_group" "this" {
  vpc_id = module.vpc.vpc_id

  name        = "bastion-host"
  description = "Securing the bastion host"
}

resource "aws_iam_role" "access_bastion" {
  name        = "connect-bastion"
  description = "Role used to connect to the bastion instance."

  assume_role_policy = jsonencode({
    Version = "2012-10-17",
    Statement = [
      {
        Effect    = "Allow",
        Action    = "sts:AssumeRole",
        Principal = { "AWS" : module.bastion_user.iam_user_arn }
    }]
  })
}

module "bastion_host" {
  # other stuff here …
  connect_bastion_role_name = aws_iam_role.access_bastion.name

  security_group_id     = aws_security_group.this.id
}
```